refactor: addContent 폴더 상위로 이동
