+++ src/components/addContent/index.js
@@ -1,6 +1,6 @@
 import AddForm from './addForm/index.js';
 import CurrentProductTable from './currentProductTable/index.js';
-import createTemplate from '../../utils/createTemplate.js';
+import createTemplate from '../utils/createTemplate.js';
 import { HEADING_ADD, HEADING_CURRENT_PRODUCT } from './const.js';
 
 const createAddFormContainer = () =>
